# docs/custom/headless-openapi.yaml
openapi: 3.0.3
info:
  title: Activepieces Headless API
  description: |
    Internal API for using Activepieces as a headless workflow automation service.

    **Authentication:** All requests require an `x-internal-api-key` header matching
    the `AP_INTERNAL_API_KEY` environment variable configured on the server. This is
    a backend-to-backend API — Activepieces must NOT be publicly exposed.

    **Base URL:** `http://<activepieces-host>:<port>/api/v1`
  version: 1.0.0
  contact:
    name: Internal
servers:
  - url: http://localhost:4200/api/v1
    description: Local development (via Vite proxy)
  - url: http://localhost:3000/api/v1
    description: Local development (direct API)
security:
  - InternalApiKey: []
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
      description: Shared secret set via `AP_INTERNAL_API_KEY` environment variable.
  schemas:
...
    AppConnection:
      type: object
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        pieceName:
          type: string
        displayName:
          type: string
        externalId:
          type: string
        projectIds:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [SECRET_TEXT, OAUTH2, PLATFORM_OAUTH2, CLOUD_OAUTH2, CUSTOM_AUTH, BASIC_AUTH, NO_AUTH]
        status:
          type: string
          enum: [ACTIVE, ERROR, MISSING]
      description: Sensitive credential values are redacted in list/get responses.
...
    SeekPage:
      type: object
      properties:
        data:
          type: array
          items: {}
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
...
paths:
...
  /app-connections:
    get:
      tags: [App Connections]
      summary: List connections in a project
      description: Returns connections with sensitive values redacted.
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: pieceName
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of connections
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SeekPage'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AppConnection'

    post:
      tags: [App Connections]
      summary: Create or update a connection (upsert)
      description: |
        Creates a new connection or updates an existing one by `externalId`.
        For internal/headless requests, engine validation is skipped.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, pieceName, displayName, type, value]
              properties:
                projectId:
                  type: string
                pieceName:
                  type: string
                  example: '@activepieces/piece-slack'
                displayName:
                  type: string
                  example: My Slack Connection
                externalId:
                  type: string
                  description: Your external identifier for this connection. Used in flow definitions as `{{connections['externalId']}}`.
                type:
                  type: string
                  enum: [SECRET_TEXT, OAUTH2, PLATFORM_OAUTH2, CLOUD_OAUTH2, CUSTOM_AUTH, BASIC_AUTH, NO_AUTH]
                value:
                  type: object
                  description: |
                    Connection credentials — structure varies by `type`:
                    - **SECRET_TEXT:** `{ type: "SECRET_TEXT", secret_text: "..." }`
                    - **OAUTH2:** `{ type: "OAUTH2", client_id, client_secret, code, scope, redirect_url, authorization_method? }`
                    - **BASIC_AUTH:** `{ type: "BASIC_AUTH", username, password }`
                    - **CUSTOM_AUTH:** `{ type: "CUSTOM_AUTH", props: { key: value, ... } }`
                    - **NO_AUTH:** `{ type: "NO_AUTH" }`
      responses:
        '201':
          description: Connection created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppConnection'
  /app-connections/{id}:
    delete:
      tags: [App Connections]
      summary: Delete a connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Connection deleted
  /app-connections/oauth2/authorization-url:
    post:
      tags: [App Connections]
      summary: Build OAuth2 authorization URL
      description: |
        Generates the OAuth2 authorization URL for a piece. Your frontend
        should open this URL in a popup/redirect. After the user authorizes,
        exchange the returned `code` via the connection upsert endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pieceName, clientId, redirectUrl]
              properties:
                pieceName:
                  type: string
                  example: '@activepieces/piece-notion'
                pieceVersion:
                  type: string
                clientId:
                  type: string
                redirectUrl:
                  type: string
                  example: http://localhost:5173/oauth/callback
                props:
                  type: object
                  description: Extra properties required by some pieces (e.g. scopes).
      responses:
        '200':
          description: Authorization URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizationUrl:
                    type: string
                    description: URL to redirect the user to for OAuth consent.
                  codeVerifier:
                    type: string
                    description: PKCE code verifier (if the piece uses PKCE).

  # ── Pieces ────────────────────────────────────────────────
...